Move blocks folder inside StreamField folder
