+++ components/StreamField/types.tsx
@@ -1,5 +1,5 @@
 import StreamFieldBlock from '../types';
-import type { Topic } from '../blocks/topics_block';
+import type { Topic } from './blocks/topics_block';
 
 export type StreamFieldProps = {
     body: Array<StreamFieldBlock>;
